Refactor to MVVM
