+++ MyApp/app/src/main/AndroidManifest.xml
@@ -10,16 +10,15 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.MyApp">
         <activity
-            android:name=".MainActivity"
+            android:name=".view.activity.MainActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <service android:name=".ContactService" />
         <receiver
-            android:name=".BirthdayAlertReceiver"
+            android:name=".viewmodel.BirthdayAlertReceiver"
             android:exported="true"
             android:enabled="true">
             <intent-filter>
