--- MyApp/app/src/main/java/com/gmail/dzutsevasabina/interfaces/ButtonClickListener.kt
@@ -1,8 +1,0 @@
-package com.gmail.dzutsevasabina.interfaces
-
-import android.widget.Button
-import com.gmail.dzutsevasabina.model.DetailedContact
-
-interface ButtonClickListener {
-    fun onButtonClick(button: Button, contact: DetailedContact)
-}
