--- MyApp/app/src/main/java/com/gmail/dzutsevasabina/interfaces/ServiceBinder.kt
@@ -1,7 +1,0 @@
-package com.gmail.dzutsevasabina.interfaces
-
-import com.gmail.dzutsevasabina.ContactService
-
-interface ServiceBinder {
-    fun getService(): ContactService?
-}
