+++ MyApp/app/src/main/res/layout/activity_main.xml
@@ -5,7 +5,7 @@
     android:layout_height="match_parent"
     android:orientation="vertical"
     android:background="@color/pale_yellow"
-    tools:context=".MainActivity">
+    tools:context=".view.activity.MainActivity">
 
     <LinearLayout
         android:id="@+id/fragments_container"
